refactor: Abolish AppProvider
