+++ src/features/request_submission/RequestSubmissionProvider.tsx
@@ -22,7 +22,7 @@
 // THE SOFTWARE.
 //
 
-import { ProviderProps } from "../../providers/ProviderProps"
+import { ProviderProps } from "../../utils/ProviderProps"
 import { LogicProvider } from "../../utils/LogicProvider"
 import { RequestSubmissionContext } from "./RequestSubmissionContext"
 import { FirestoreRequestSubmissionRepository } from "./RequestSubmissionRepository"
